Converted light/dark mode switch to a component
