+++ src/App.jsx
@@ -1,19 +1,17 @@
-import React, { useReducer, useState } from 'react';
+import React, { useReducer } from 'react';
 import styles from './App.module.scss';
 import {
   GlobalContext,
   globalState,
   globalReducer,
 } from './context/reducers/globalReducer';
-import { toggleTheme } from './context/actions/globalActions';
 import ProjectLoader from './pages/ProjectLoader/ProjectLoader.jsx';
 import NavBar from './components/NavBar/NavBar';
+import ModeSwitch from './components/ModeSwitch/ModeSwitch';
 import LeftPanel from './pages/LeftPanel/LeftPanel';
 import RightPanel from './pages/RightPanel/RightPanel';
 import FileDirectory from './components/FileDirectory/FileDirectory';
 import { CSSTransition } from 'react-transition-group';
-import { Switch } from '@mui/material';
-import { BiSun, BiMoon } from 'react-icons/bi';
 import { ThemeProvider, StyledEngineProvider, createTheme } from '@mui/material/styles';
 
 const theme = createTheme();
@@ -21,34 +19,16 @@ const theme = createTheme();
 const App = () => {
   const [global, dispatchToGlobal] = useReducer(globalReducer, globalState);
 
-  const changeTheme = () => {
-    localStorage.setItem('theme', global.theme === 'light' ? 'dark' : 'light');
-    dispatchToGlobal(toggleTheme());
-  };
-
+  // render project loader page that enables user to pick a project
   if (!global.isProjectLoaded) {
     return (
       <div>
-
         {/* pass global state and dispatch function as prop to context provider for child components */}
         <GlobalContext.Provider value={[global, dispatchToGlobal]}>
           <ThemeProvider theme={theme} >
             <StyledEngineProvider injectFirst>
                 <div id={styles.toggle}>
-                  <div id={styles.icon}>
-                    <span title="Dark Mode">
-                      <BiMoon size={'1.5rem'} />
-                    </span>
-                    <span title="Change theme">
-                      <Switch
-                        checked={global.theme === 'light' ? true : false}
-                        onChange={changeTheme}
-                      />
-                    </span>
-                    <span title="Light Mode">
-                      <BiSun size={'1.5rem'} />
-                    </span>
-                  </div>
+                  <ModeSwitch/>
                 </div>
                 <ProjectLoader />
             </StyledEngineProvider>
@@ -57,7 +37,7 @@ const App = () => {
       </div>
     );
   }
-  
+  // after user picks project, load page that tells user to choose a test
   if (global.testCase === '') {
     return (
       <div>
@@ -96,9 +76,7 @@ const App = () => {
      *
      * We access the value that we gave to the Provider through useContext
      */
-    <div
-      id={styles.app}
-    >
+    <div id={styles.app}>
       <GlobalContext.Provider value={[global, dispatchToGlobal]}>
         <ThemeProvider theme={theme} >
           <StyledEngineProvider injectFirst>
