+++ src/App.module.scss
@@ -98,25 +98,3 @@
   right: 10px;
   z-index: 3000;
 }
-
-#icon {
-  flex: 0 0 auto;
-  display: flex;
-  align-items: center;
-
-  span{
-    svg {
-      path {
-        fill: $eggshell; 
-      }
-    }
-    
-  }
-  #activeEffect {
-    svg {
-      path {
-        fill: var(--icon-fill);
-      }
-    }
-  }
-}
