+++ src/pages/ProjectLoader/ProjectLoader.jsx
@@ -11,7 +11,7 @@ const { ipcRenderer } = require('electron');
 // const remote = require('@electron/remote/main')
 
 function ProjectLoader() {
-  const [{ idFileDirectoryOpen, theme }, dispatchToGlobal] = useContext(GlobalContext);
+  const [{ isFileDirectoryOpen, theme }, dispatchToGlobal] = useContext(GlobalContext);
 
   const [username, setUsername] = useState('');
   const [password, setPassword] = useState('');
